+++ backend/jest/setup.loadEnv.ts
@@ -1,0 +1,3 @@
+import { loadEnv } from "../src/env";
+
+loadEnv();
