+++ backend/package.json
@@ -14,15 +14,16 @@
     "start": "node ./build/index.js",
     "build": "tsc",
     "dev": "tsx watch src/index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test:unit": "jest",
     "test:lint:eslint": "eslint --ext .ts,.tsx,.js,.jsx,.cjs,.mjs,.json,.yml,.yaml --max-warnings 0 .",
     "update": "npx npm-check-updates"
   },
   "devDependencies": {
     "@eslint-community/eslint-plugin-eslint-comments": "^4.4.1",
+    "@types/jest": "^29.5.14",
+    "@types/nodemailer": "^6.4.17",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "@types/nodemailer": "^6.4.17",
     "eslint": "^8.24.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-config-standard": "^17.1.0",
@@ -38,6 +39,8 @@
     "eslint-plugin-react-refresh": "^0.4.18",
     "eslint-plugin-security": "^3.0.1",
     "eslint-plugin-yml": "^1.14.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.6",
     "tsx": "^4.19.3",
     "typescript": "^5.7.3"
   },
@@ -46,5 +49,19 @@
     "fastify": "^5.2.1",
     "nodemailer": "^6.10.0",
     "ts-dotenv": "^0.9.1"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "setupFiles": ["./jest/setup.loadEnv.ts"],
+    "collectCoverage": true,
+    "coverageThreshold": {
+      "global": {
+        "branches": 100,
+        "functions": 100,
+        "lines": 100,
+        "statements": 100
+      }
+    }
   }
 }
